requirements.txt
django
djangorestframework
djangorestframework-simplejwt
razorpay
mysqlclient